backend-chat/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=8.1
django-cors-headers>=4.0
Pillow>=10.0
gunicorn
